requirements.txt
# Video Ad Placement Service Dependencies

# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1

# File Operations
aiofiles==23.2.1

# HTTP Client for testing
httpx==0.25.2

# Data Processing and ML
numpy==1.24.4
opencv-python==4.8.1.78
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.36.2

# Async Processing
asyncio-throttle==1.0.2
celery==5.3.4

# Monitoring and Metrics
psutil==5.9.6
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Optional Dependencies (for production)
# gunicorn==21.2.0
# sentry-sdk[fastapi]==1.38.0
# structlog==23.2.0 